refactor: use extend annotation reader
