+++ src/AnnotationsReader.ts
@@ -1,6 +1,1 @@
-import ts from "typescript";
-import { Annotations } from "./Type/AnnotatedType";
-
-export interface AnnotationsReader {
-    getAnnotations(node: ts.Node): Annotations | undefined;
-}
+export { AnnotationsReader } from "./Interfaces/AnnotationsReader";
