+++ src/Interfaces/AnnotationsReader.ts
@@ -1,6 +1,8 @@
 import ts from "typescript";
-import { Annotations } from "../Type/AnnotatedType";
+import type { Annotations } from "../Type/AnnotatedType";
+import type { BaseType } from "../Type/BaseType";
 
 export interface AnnotationsReader {
     getAnnotations(node: ts.Node): Annotations | undefined;
+    getAnnotationsFromType?(type: BaseType): Annotations | undefined;
 }
